Added a flag for removing instances

+++ kill-instances.sh
@@ -2,9 +2,11 @@
 
 REGION="us-east-1"
 
-attempted_instanceids=`cat ${BUILD_NUMBER}_instanceIds.txt`
-aws ec2 terminate-instances --instance-ids ${attempted_instanceids[@]} \
-  --region $REGION \
-  --output text \
-  --query 'TerminatingInstances[].InstanceId'
-echo
+if [ $KILL_INSTANCES -eq "true" ] ; then
+	attempted_instanceids=`cat ${BUILD_NUMBER}_instanceIds.txt`
+	aws ec2 terminate-instances --instance-ids ${attempted_instanceids[@]} \
+	  --region $REGION \
+	  --output text \
+	  --query 'TerminatingInstances[].InstanceId'
+	echo
+fi
